face-service/requirements.txt
flask==3.1.2
flask-cors>=5.0.0
face_recognition==1.3.0
opencv-python==4.12.0.88
numpy==2.2.6
pillow==12.1.0
